add snackbar and layout group1_item_movie-land
