+++ app/src/main/java/site/madcat/movielibrary/domain/GetJSONMovieInterface.kt
@@ -2,8 +2,8 @@ package site.madcat.movielibrary.domain
 
 import androidx.annotation.WorkerThread
 
-interface GetMovieRepoInterface {
+interface GetJSONMovieInterface {
     @WorkerThread
     fun getMovieSync(path:String): List<MovieEntity> //или ReturnPackage
-    fun getMovieAsync(path:String, callback:(List<MovieEntity>)->Unit)//или ReturnPackage
+
 }
