+++ app/src/main/java/site/madcat/movielibrary/ui/activity/MovieActivity.kt
@@ -16,15 +16,22 @@ import site.madcat.movielibrary.ui.favoritesFragment.FavoritesFragment
 import site.madcat.movielibrary.ui.homeFragment.HomeFragment
 import site.madcat.movielibrary.ui.raitingFragment.RaitingFragment
 import com.google.android.material.snackbar.Snackbar
+import site.madcat.movielibrary.app
+import site.madcat.movielibrary.domain.GetJSONMovieInterface
 
 
 class MovieActivity : AppCompatActivity(), HomeFragment.Controller {
+
+
     private lateinit var binding: ActivityMainBinding
     private var movieActivityPresenter=MovieActivityPresenter()
     private var fragmentManager: FragmentManager=supportFragmentManager
     private lateinit var bottomNavigationItemView: BottomNavigationView
     private lateinit var baseSnackView: View
+    val repository: GetJSONMovieInterface by lazy { app.repository }
 
+    val urlPath: String=
+        "https://api.themoviedb.org/3/discover/movie?&sort_by=popularity.desc&api_key=b46aa2f69329d4b3b5e8d2e1ea6b7886"
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -37,6 +44,13 @@ class MovieActivity : AppCompatActivity(), HomeFragment.Controller {
             loadFragment(HomeFragment())
         }
         initNavigation()
+
+        Thread {
+            repository.getMovieSync(urlPath)
+        }.start()
+
+
+       // movieActivityPresenter.getMovie()
     }
 
 
@@ -90,6 +104,7 @@ class MovieActivity : AppCompatActivity(), HomeFragment.Controller {
     override fun loadMovie(movie: MovieEntity?) {
         setDetailMovieFragment(movie)
     }
+
 
 
     fun View.showSnackBar(
