+++ app/src/main/java/site/madcat/movielibrary/ui/activity/MovieActivityContract.kt
@@ -3,11 +3,9 @@ package site.madcat.movielibrary.ui.activity
 import site.madcat.movielibrary.domain.MovieEntity
 
 class MovieActivityContract {
-
     interface MovieActivityInterface {
-
         fun onAttach(view: MovieActivity)
         fun onDetach()
-        fun fillRepository(movie:MovieEntity)
+fun getMovie()
     }
 }
