+++ app/src/main/java/site/madcat/movielibrary/ui/homeFragment/HomeFragmentPresenter.kt
@@ -4,8 +4,9 @@ package site.madcat.movielibrary.ui.homeFragment
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import site.madcat.movielibrary.App
+import site.madcat.movielibrary.app
+import site.madcat.movielibrary.domain.GetJSONMovieInterface
 import site.madcat.movielibrary.domain.MovieEntity
-import site.madcat.movielibrary.domain.MovieRepository
 import site.madcat.movielibrary.ui.home.HomeAdapter
 
 
@@ -13,12 +14,12 @@ class HomeFragmentPresenter() : ViewModel(), HomeFragmentContract.HomeFragmentIn
     HomeAdapter.IItemClickListener {
     override val respond=MutableLiveData<String>()
     private var view: HomeFragment?=null
-    lateinit var repository: MovieRepository
+ lateinit var  repository: GetJSONMovieInterface
 
 
     override fun onAttach(view: HomeFragment) {
         this.view=view
-      repository=(view.requireActivity().application as App).repository
+     repository=(view.requireActivity().application as App).repository
     }
 
     override fun onDetach() {
@@ -35,3 +36,4 @@ class HomeFragmentPresenter() : ViewModel(), HomeFragmentContract.HomeFragmentIn
 
 
 }
+
